sharp lab4 +lab6

+++ 2 cem/C SHARP/Lab_4/lab4/libmath/mathlib.cpp
@@ -1,7 +1,7 @@
 #include<stdio.h>
 
 extern "C" {
-  __declspec(dllexport) int Add(double a, double b)
+  __declspec(dllexport) int Add(int a, int b)
   {
     return a + b;
   }
